Add channel_timeout parameter to FortiOS
A new channel_timeout parameter was added to the FortiOS constructor.

It allows one to specify a time in seconds to wait for blocking commands that were executed over the SSH channel.

Until now this value was hardcoded to 5 seconds. However this is too short if the Fortigate firewall has a large configuration. Therefore it can be overwritten now if necessary.