Changed behaviour how FortiOS class reads data from SSH channel
The way how the execute_command function of the FortiOS class reads data from the SSH channel has been rewritten.
The old implementation was prone to timeout issues when the configuration of the firewall is espcially large (for example if it contains around 60.000 policies).
The new implementation only reads from the channel if data is available and writes it to an interm BytesIO Buffer. This should avoid timeouts of the blocking receive calls.