Add support for Multi Line Values
Some fields such as the comments of firewall policies can contain new lines. This means that the config value takes up multiple lines in the configuration.

Currently the library was not possible to detect such cases which led to errors and incomplete or even invalid configurations.

This commit changes the implementation of the parse_config_output function such that muli line values are detected and handled correctly.